Sync puppet module rsync to v0.3.0 from upstream
v0.3.0 sha1: d0707c4ae3674b63c2193c0012ceb9fa90ed41b6
Repo link: https://github.com/puppetlabs/puppetlabs-rsync
Implements: blueprint merge-openstack-puppet-modules

+++ deployment/puppet/rsync/manifests/get.pp
@@ -26,15 +26,20 @@
 #
 define rsync::get (
   $source,
-  $path = undef,
-  $user = undef,
-  $purge = undef,
-  $exclude = undef,
-  $keyfile = undef,
-  $timeout = '900'
+  $path       = $name,
+  $user       = undef,
+  $purge      = undef,
+  $recursive  = undef,
+  $links      = undef,
+  $hardlinks  = undef,
+  $copylinks  = undef,
+  $times      = undef,
+  $include    = undef,
+  $exclude    = undef,
+  $keyfile    = undef,
+  $timeout    = '900',
+  $execuser   = 'root',
 ) {
-
-  include rsync
 
   if $keyfile {
     $Mykeyfile = $keyfile
@@ -47,26 +52,52 @@ define rsync::get (
   }
 
   if $purge {
-    $MyPurge = '--delete'
+    $MyPurge = ' --delete'
   }
 
+  # Not currently correct, there can be multiple --exclude arguments
   if $exclude {
-    $MyExclude = "--exclude=${exclude}"
+    $MyExclude = " --exclude=${exclude}"
   }
 
-  if $path {
-    $MyPath = $path
-  } else {
-    $MyPath = $name
+  # Not currently correct, there can be multiple --include arguments
+  if $include {
+    $MyInclude = " --include=${include}"
   }
 
-  $rsync_options = "-a ${MyPurge} ${MyExclude} ${MyUser}${source} ${MyPath}"
+  if $recursive {
+    $MyRecursive = ' -r'
+  }
+
+  if $links {
+    $MyLinks = ' --links'
+  }
+
+  if $hardlinks {
+    $MyHardLinks = ' --hard-links'
+  }
+
+  if $copylinks {
+    $MyCopyLinks = ' --copy-links'
+  }
+
+  if $times {
+    $MyTimes = ' --times'
+  }
+
+  $rsync_options = "-a${MyPurge}${MyExclude}${MyInclude}${MyLinks}${MyHardLinks}${MyCopyLinks}${MyTimes}${MyRecursive} ${MyUser}${source} ${path}"
 
   exec { "rsync ${name}":
     command => "rsync -q ${rsync_options}",
-    path    => [ '/bin', '/usr/bin' ],
+    path    => [ '/bin', '/usr/bin', '/usr/local/bin' ],
+    user => $execuser,
+    # perform a dry-run to determine if anything needs to be updated
+    # this ensures that we only actually create a Puppet event if something needs to
+    # be updated
+    # TODO - it may make senes to do an actual run here (instead of a dry run)
+    #        and relace the command with an echo statement or something to ensure
+    #        that we only actually run rsync once
+    onlyif  => "test `rsync --dry-run --itemize-changes ${rsync_options} | wc -l` -gt 0",
     timeout => $timeout,
   }
-
-  Package['rsync'] -> Exec["rsync ${name}"]
 }
